backend/requirements.txt
# Django and Django REST Framework
Django==5.2.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0

# CORS headers for cross-origin requests
django-cors-headers==4.3.1

# Database adapters
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Development tools
pylint==3.0.3
black==24.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0  # For serving static files

# Security
python-dotenv==1.0.0  # For environment variables
